socket: retype the public socket-handle API surface to KlSocketHandle (PAL Phase 6a-2)
Widen every public struct field and function signature that carries a socket
DESCRIPTOR from int to KlSocketHandle (intptr_t), so a Winsock SOCKET (unsigned,
pointer-width) is held losslessly. POSIX behavior-preserving; MinGW-validated.

Retyped surface:
- Struct fields: KlConn.fd, KlUdp.fd, KlWatcher.fd, KlResponse.conn_fd,
  KlConfig.listen_fd + KlServer.listen_fd, KlClientPoolConn.fd (x2).
- Callback/vtable types: KlWatcherFn; the KlTls vtable (handshake/read/write/
  shutdown); the KlFileIO vtable (sock_fd on submit/cancel).
- APIs: kl_event_add/mod/del, kl_watcher_add/mod/del/rearm, kl_conn_acquire,
  kl_response_file, kl_udp_fd, kl_udp_server_fd, kl_peer_cred_fd.
- Every consumer forced by -Werror: watcher callbacks + event backends (epoll/
  kqueue/poll/io_uring) + TLS/file_io impls, across src, tests, and examples.

Stays int (genuinely a different handle on Windows): KlEventLoop.fd (backend-
internal epoll/kqueue fd), KlResponse.file_fd / KlFileIO file_fd / sendfile in_fd
(FILE descriptors, CRT fds on Windows).

Behavior-preserving on POSIX: KlSocketHandle == intptr_t, so an int fd round-trips
losslessly and -1 stays invalid; no -Wconversion, so consumer locals that store
`int fd = kl_sock_accept(...)` narrow silently -- those variable conversions +
the fd<0 -> kl_handle_valid check style ride with 6a-3 (validated by the MinGW
build's -Wconversion, where the .c actually cross-compiles).

Validated: macOS + Linux glibc (epoll + poll + io_uring BUILD) + musl/Alpine full
suite, smoke, ASan/UBSan, scan-build (no bugs), cppcheck 2.13, and a MinGW compile
of the retyped public headers under winsock2.h (the handle-bearing surface is
SOCKET-compatible). The io_uring build leg caught a real regression -- the Linux-
only file_io_iouring.c still assigned int-typed submit/cancel impls to the now-
KlSocketHandle KlFileIO vtable -- fixed here. (io_uring runtime tests don't run
in Docker-on-macOS -- confirmed environmental, clean base fails identically; CI's
runner is the runtime gate.)

+++ examples/connection_pool.c
@@ -43,7 +43,7 @@ int main(void) {
     for (int i = 0; i < capacity; i++) {
         conns[i] = kl_conn_acquire(&pool, 100 + i);
         printf("  acquired fd=%d, active=%d, state=%u\n",
-               conns[i] ? conns[i]->fd : -1,
+               conns[i] ? (int)conns[i]->fd : -1,
                pool.active_count,
                conns[i] ? (unsigned)conns[i]->state : 0);
     }
@@ -62,7 +62,7 @@ int main(void) {
 
     KlConn *reused = kl_conn_acquire(&pool, 201);
     printf("  re-acquired fd=%d, active=%d\n",
-           reused ? reused->fd : -1, pool.active_count);
+           reused ? (int)reused->fd : -1, pool.active_count);
 
     /* Timing */
     printf("\n--- Monotonic clock ---\n");
